Port this change from Nancy.

This also fixes a bug with multiple roots, where copying files would only
work from the first root.

Also update README.nancy.md slightly to improve the language, also copying
from Nancy.

+++ src/index.ts
@@ -1,12 +1,9 @@
-import fs from 'fs-extra'
+import fs from 'fs'
+import fsExtra from 'fs-extra' // See https://github.com/jprichardson/node-fs-extra/issues/919
 import path from 'path'
 import Debug from 'debug'
 import assert from 'assert'
 import execa from 'execa'
-import realFs from 'fs'
-import {link} from 'linkfs'
-import {IUnionFs, Union} from 'unionfs'
-import {IFS} from 'unionfs/lib/fs'
 import slimdom from 'slimdom'
 import {sync as parseXML} from 'slimdom-sax-parser'
 import formatXML from 'xml-formatter'
@@ -21,19 +18,33 @@ export function stripPathPrefix(s: string, prefix: string): string {
   if (s.startsWith(prefix + path.sep)) {
     return path.join(s.slice(prefix.length + path.sep.length))
   }
-  if (s === prefix) {
-    return ''
-  }
-  return s
+  return s === prefix ? '' : s
 }
 
-// Merge input directories, left as highest-priority
-export function unionFs(dirs: string[]): IUnionFs {
-  const ufs = new Union();
-  for (const dir of dirs.slice(1).reverse()) {
-    ufs.use(link(fs, [dirs[0], dir]))
+function isExecutable(file: string): boolean {
+  try {
+    fs.accessSync(file, fs.constants.X_OK)
+    return true
+  } catch {
+    return false
   }
-  return ufs.use(realFs)
+}
+
+// FIXME: `throwIfNoEntry` is missing in TypeScript types for Node 14:
+// https://github.com/DefinitelyTyped/DefinitelyTyped/discussions/55786
+function statSync(file: string): fs.Stats {
+  return (fs as any).statSync(file, {throwIfNoEntry: false})
+}
+
+type FullDirent = fs.Dirent & {path: string}
+type File = string
+type Directory = FullDirent[]
+type Dirent = File | Directory | undefined
+function isFile(object: Dirent): object is File {
+  return typeof object === 'string'
+}
+function isDirectory(object: Dirent): object is Directory {
+  return Array.isArray(object)
 }
 
 const ruth = 'https://github.com/rrthomas/ruth/raw/main/ruth.dtd'
@@ -46,21 +57,33 @@ const xQueryOptions: Options = {
   debug: process.env.DEBUG !== undefined,
 }
 
-export class Expander {
-  private absInput: string
+function loadModule(file: string) {
+  const module = fs.readFileSync(file, 'utf-8')
+  registerXQueryModule(module)
+  const matches = /^\s*module\s+namespace\s+([^= ]+)\s*=\s*"([^"]+)"\s*;\s*$/m.exec(module)
+  // If there was no module declaration, registerXQueryModule would have errored,
+  // so if 'matches' is null, the parser (regex above) needs improving.
+  assert(matches !== null)
+  if (xQueryOptions.moduleImports === undefined) {
+    xQueryOptions.moduleImports = {}
+  }
+  const prefix = matches[1]
+  const url = matches[2]
+  xQueryOptions.moduleImports[prefix] = url
+  debug(`registered prefix ${prefix} for URL ${url}`)
+}
 
+export class Expander {
   private xtree: slimdom.Document
 
   constructor(
-    private input: string,
-    private inputFs: IFS = realFs,
+    private inputs: string[],
     private xmlExtensions: string[] = [],
     private maxIterations = 8,
   ) {
     this.xmlExtensions = this.xmlExtensions.concat('.xml', '.xhtml')
-    this.absInput = path.resolve(input)
-    this.loadModule(path.join(__dirname, 'ruth.xq'))
-    this.xtree = this.dirTreeToXML(input)
+    loadModule(path.join(__dirname, 'ruth.xq'))
+    this.xtree = this.dirTreeToXML('')
     // FIXME: The next line only works once, so can only use Expander once.
     // See https://github.com/FontoXML/fontoxpath/issues/406
     registerCustomXPathFunction(
@@ -77,60 +100,84 @@ export class Expander {
         ) as slimdom.Node[]
       },
     )
+    registerCustomXPathFunction(
+      {localName: 'absolute-path', namespaceURI: ruth},
+      ['xs:string'], 'xs:string',
+      (_, path: string): string => {
+        debug(`ruth:absolute-path(${path})`)
+        const dirent = this.findObject(path)
+        if (isFile(dirent)) {
+          return dirent
+        }
+        throw new Error(`${path} is not a file`)
+      },
+    )
+  }
+
+  // Find the first file or directory with path `object` in the input tree,
+  // scanning the roots from left to right.
+  // If the result is a file, return its file system path.
+  // If the result is a directory, return its contents as a map from tree
+  // paths to file system `fs.Dirent`s, obtained by similarly scanning the
+  // tree from left to right.
+  // If something neither a file nor directory is found, raise an error.
+  // If no result is found, return `undefined`.
+  private findObject(object: string): Dirent {
+    const dirs = []
+    for (const root of this.inputs) {
+      const stats = statSync(root)
+      if (stats !== undefined && (statSync(root).isDirectory() || object === '')) {
+        const objectPath = path.join(root, object)
+        const stats = statSync(objectPath)
+        if (stats !== undefined) {
+          if (stats.isFile()) {
+            return objectPath
+          }
+          if (stats.isDirectory()) {
+            dirs.push(objectPath)
+          } else {
+            throw new Error(`${objectPath} is not a file or directory`)
+          }
+        }
+      }
+    }
+    const dirents: Directory = []
+    for (const dir of dirs.reverse()) {
+      for (const dirent of fs.readdirSync(dir, {withFileTypes: true})) {
+        const fullDirent: FullDirent = dirent as FullDirent
+        fullDirent.path = path.join(dir, dirent.name)
+        dirents.push(fullDirent)
+      }
+    }
+    return dirs.length > 0 ? dirents : undefined
   }
 
   private static templateRegex = /\.ruth([0-9])*(?=\.[^.]+$|$)/
 
   private static noCopyRegex = /\.in(?=\.[^.]+$|$)/
 
-  private isExecutable(file: string): boolean {
-    try {
-      this.inputFs.accessSync(file, fs.constants.X_OK)
-      return true
-    } catch {
-      return false
-    }
-  }
-
-  private loadModule(file: string) {
-    const module = this.inputFs.readFileSync(file, 'utf-8')
-    registerXQueryModule(module)
-    const matches = /^\s*module\s+namespace\s+([^= ]+)\s*=\s*"([^"]+)"\s*;\s*$/m.exec(module)
-    // If there was no module declaration, registerXQueryModule would have errored,
-    // so if 'matches' is null, the parser (regex above) needs improving.
-    assert(matches !== null)
-    if (xQueryOptions.moduleImports === undefined) {
-      xQueryOptions.moduleImports = {}
-    }
-    const prefix = matches[1]
-    const url = matches[2]
-    xQueryOptions.moduleImports[prefix] = url
-    debug(`registered prefix ${prefix} for URL ${url}`)
-  }
-
   private dirTreeToXML(root: string) {
     const xtree = new slimdom.Document()
     const objToNode = (obj: string) => {
-      const stats = this.inputFs.statSync(obj)
+      const realObj = this.findObject(obj)
       const parsedPath = path.parse(obj)
       let elem: slimdom.Element
       debug(`dirTreeToXML: considering ${obj}`)
-      if (stats.isDirectory()) {
+      if (isDirectory(realObj)) {
         debug('processing directory')
         elem = xtree.createElementNS(dirtree, 'directory')
-        const dir = this.inputFs.readdirSync(obj, {withFileTypes: true})
-          .filter((dirent) => dirent.name[0] !== '.')
+        const dir = realObj.filter((dirent) => dirent.name[0] !== '.')
         const dirs = dir.filter((dirent) => dirent.isDirectory()).sort()
         const files = dir.filter((dirent) => dirent.isFile() || dirent.isSymbolicLink()).sort()
         dirs.forEach((dirent) => elem.appendChild(objToNode(path.join(obj, dirent.name))))
         files.forEach((dirent) => elem.appendChild(objToNode(path.join(obj, dirent.name))))
-      } else if (stats.isFile() || stats.isSymbolicLink()) {
+      } else if (isFile(realObj)) {
         debug('processing file')
-        if (this.isExecutable(obj)) {
+        if (isExecutable(realObj)) {
           debug('creating XQuery function from executable')
           const localName = (/^[^.]*/.exec(parsedPath.name) as string[])[0]
           const exec = (_: any, args: string[], input?: string): string => execa.sync(
-            path.join(this.absInput, stripPathPrefix(obj, this.input)), args, {input},
+            realObj, args, {input},
           ).stdout
           registerCustomXPathFunction(
             {localName, namespaceURI: ruth}, ['xs:string*'], 'xs:string', exec,
@@ -141,7 +188,7 @@ export class Expander {
           elem = xtree.createElementNS(dirtree, 'file')
         } else if (this.xmlExtensions.includes(parsedPath.ext)) {
           debug('reading as XML')
-          const text = this.inputFs.readFileSync(obj, 'utf-8')
+          const text = fs.readFileSync(realObj, 'utf-8')
           const wrappedText = `<dirtree:file>${text}</dirtree:file>`
           let doc
           try {
@@ -155,14 +202,14 @@ export class Expander {
           debug('not reading as XML')
           if (/.xq[lmy]?/.test(parsedPath.ext)) {
             debug('reading as XQuery module')
-            this.loadModule(obj)
+            loadModule(realObj)
           }
           elem = xtree.createElementNS(dirtree, 'file')
         }
       } else {
-        throw new Error(`'${obj}' is not a directory or file`)
+        throw new Error(`'${obj}' is not a file or directory`)
       }
-      elem.setAttributeNS(dirtree, 'path', stripPathPrefix(obj, this.input))
+      elem.setAttributeNS(dirtree, 'path', obj)
       elem.setAttributeNS(dirtree, 'name', parsedPath.base)
       return elem
     }
@@ -176,11 +223,10 @@ export class Expander {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   xQueryVariables: {[id: string]: any} = {
     // These variables are injected into the ruth namespace in lib/ruth.xq.
-    root: this.input,
   }
 
   private index(filePath: string): slimdom.Element {
-    const components = [path.basename(this.input)]
+    const components = ['']
     if (filePath !== '') {
       components.push(...filePath.split(path.sep))
     }
@@ -210,7 +256,7 @@ export class Expander {
       const outputPath = path.join(outputDir, stripPathPrefix(obj, buildPath))
       if (elem.namespaceURI === dirtree && elem.localName === 'directory') {
         debug('Expanding directory')
-        fs.emptyDirSync(outputPath)
+        fsExtra.ensureDirSync(outputPath)
         elem.children.filter((child) => child.tagName !== 'directory').forEach(expandElement)
         elem.children.filter((child) => child.tagName === 'directory').forEach(expandElement)
       } else {
@@ -263,7 +309,10 @@ export class Expander {
         elem.replaceWith(expandedElem)
         fs.writeFileSync(outputPath, expandedElem.innerHTML)
       } else if (!Expander.noCopyRegex.exec(obj)) {
-        const objFullPath = path.join(this.input, obj)
+        const objFullPath = this.findObject(obj)
+        if (!isFile(objFullPath)) {
+          throw new Error(`${obj} is not a file`)
+        }
         fs.copyFileSync(objFullPath, outputPath)
       }
     }
